@@ -10,24 +10,16 @@ namespace Chinook.API.Controllers;
 [Route("api/[controller]")]
 [ApiController]
 [EnableCors("CorsPolicy")]
-public class GenreController : ControllerBase
+public class GenreController(IChinookSupervisor chinookSupervisor, ILogger<GenreController> logger)
+    : ControllerBase
 {
-    private readonly IChinookSupervisor _chinookSupervisor;
-    private readonly ILogger<GenreController> _logger;
-
-    public GenreController(IChinookSupervisor chinookSupervisor, ILogger<GenreController> logger)
-    {
-        _chinookSupervisor = chinookSupervisor;
-        _logger = logger;
-    }
-
     [HttpGet]
     [Produces("application/json")]
     public ActionResult<List<GenreApiModel>> Get()
     {
         try
         {
-            var genres = _chinookSupervisor.GetAllGenre();
+            var genres = chinookSupervisor.GetAllGenre();
 
             if (genres.Any())
             {
@@ -40,7 +32,7 @@ public class GenreController : ControllerBase
         }
         catch (Exception ex)
         {
-            _logger.LogError($"Something went wrong inside the GenreController Get action: {ex}");
+            logger.LogError($"Something went wrong inside the GenreController Get action: {ex}");
             return StatusCode((int)HttpStatusCode.InternalServerError,
                 "Error occurred while executing Get All Genres");
         }
@@ -52,7 +44,7 @@ public class GenreController : ControllerBase
     {
         try
         {
-            var genre = _chinookSupervisor.GetGenreById(id);
+            var genre = chinookSupervisor.GetGenreById(id);
 
             if (genre != null)
             {
@@ -65,7 +57,7 @@ public class GenreController : ControllerBase
         }
         catch (Exception ex)
         {
-            _logger.LogError($"Something went wrong inside the GenreController GetById action: {ex}");
+            logger.LogError($"Something went wrong inside the GenreController GetById action: {ex}");
             return StatusCode((int)HttpStatusCode.InternalServerError,
                 "Error occurred while executing Get Genre By Id");
         }
@@ -84,17 +76,17 @@ public class GenreController : ControllerBase
             }
             else
             {
-                return Ok(_chinookSupervisor.AddGenre(input));
+                return Ok(chinookSupervisor.AddGenre(input));
             }
         }
         catch (ValidationException ex)
         {
-            _logger.LogError($"Something went wrong inside the GenreController Add Genre action: {ex}");
+            logger.LogError($"Something went wrong inside the GenreController Add Genre action: {ex}");
             return StatusCode((int)HttpStatusCode.InternalServerError, "Error occurred while executing Add Genres");
         }
         catch (Exception ex)
         {
-            _logger.LogError($"Something went wrong inside the GenreController Add Genre action: {ex}");
+            logger.LogError($"Something went wrong inside the GenreController Add Genre action: {ex}");
             return StatusCode((int)HttpStatusCode.InternalServerError, "Error occurred while executing Add Genres");
         }
     }
@@ -112,18 +104,18 @@ public class GenreController : ControllerBase
             }
             else
             {
-                return Ok(_chinookSupervisor.UpdateGenre(input));
+                return Ok(chinookSupervisor.UpdateGenre(input));
             }
         }
         catch (ValidationException ex)
         {
-            _logger.LogError($"Something went wrong inside the GenreController Update Genre action: {ex}");
+            logger.LogError($"Something went wrong inside the GenreController Update Genre action: {ex}");
             return StatusCode((int)HttpStatusCode.InternalServerError,
                 "Error occurred while executing Update Genres");
         }
         catch (Exception ex)
         {
-            _logger.LogError($"Something went wrong inside the GenreController Update Genre action: {ex}");
+            logger.LogError($"Something went wrong inside the GenreController Update Genre action: {ex}");
             return StatusCode((int)HttpStatusCode.InternalServerError,
                 "Error occurred while executing AUpdatedd Genres");
         }
@@ -134,11 +126,11 @@ public class GenreController : ControllerBase
     {
         try
         {
-            return Ok(_chinookSupervisor.DeleteGenre(id));
+            return Ok(chinookSupervisor.DeleteGenre(id));
         }
         catch (Exception ex)
         {
-            _logger.LogError($"Something went wrong inside the GenreController Delete action: {ex}");
+            logger.LogError($"Something went wrong inside the GenreController Delete action: {ex}");
             return StatusCode((int)HttpStatusCode.InternalServerError,
                 "Error occurred while executing Delete Genre");
         }
