@@ -1,20 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata;
+﻿using Microsoft.EntityFrameworkCore;
 using Chinook.Domain.Entities;
 
 namespace Chinook.Data.Data;
 
-public partial class ChinookContext : DbContext
+public partial class ChinookContext(DbContextOptions<ChinookContext> options) : DbContext(options)
 {
-    public ChinookContext(DbContextOptions<ChinookContext> options)
-        : base(options)
-    {
-    }
-
     public virtual DbSet<Album> Albums { get; set; } = null!;
-    public virtual DbSet<AlbumWithArtistName> AlbumWithArtistNames { get; set; } = null!;
     public virtual DbSet<Artist> Artists { get; set; } = null!;
     public virtual DbSet<Customer> Customers { get; set; } = null!;
     public virtual DbSet<Employee> Employees { get; set; } = null!;
@@ -42,17 +33,6 @@ public partial class ChinookContext : DbContext
                 .HasForeignKey(d => d.ArtistId)
                 .OnDelete(DeleteBehavior.ClientSetNull)
                 .HasConstraintName("FK__Album__ArtistId__276EDEB3");
-        });
-
-        modelBuilder.Entity<AlbumWithArtistName>(entity =>
-        {
-            entity.HasNoKey();
-
-            entity.ToView("AlbumWithArtistName");
-
-            entity.Property(e => e.Name).HasMaxLength(120);
-
-            entity.Property(e => e.Title).HasMaxLength(160);
         });
 
         modelBuilder.Entity<Artist>(entity =>
