@@ -5,30 +5,23 @@ using Microsoft.EntityFrameworkCore;
 
 namespace Chinook.Data.Repositories;
 
-public class CustomerRepository : ICustomerRepository
+public class CustomerRepository(ChinookContext context) : ICustomerRepository
 {
-    private readonly ChinookContext _context;
-
-    public CustomerRepository(ChinookContext context)
-    {
-        _context = context;
-    }
-
     private bool CustomerExists(int id) =>
-        _context.Customers.Any(c => c.Id == id);
+        context.Customers.Any(c => c.Id == id);
 
-    public void Dispose() => _context.Dispose();
+    public void Dispose() => context.Dispose();
 
     public List<Customer> GetAll() =>
-        _context.Customers.AsNoTrackingWithIdentityResolution().ToList();
+        context.Customers.Include(c => c.Invoices).Include(c => c.SupportRep).AsNoTrackingWithIdentityResolution().ToList();
 
     public Customer GetById(int id) =>
-        _context.Customers.Find(id);
+        context.Customers.Include(c => c.Invoices).Include(c => c.SupportRep).AsNoTrackingWithIdentityResolution().FirstOrDefault(c => c.Id == id);
 
     public Customer Add(Customer newCustomer)
     {
-        _context.Customers.Add(newCustomer);
-        _context.SaveChanges();
+        context.Customers.Add(newCustomer);
+        context.SaveChanges();
         return newCustomer;
     }
 
@@ -36,8 +29,8 @@ public class CustomerRepository : ICustomerRepository
     {
         if (!CustomerExists(customer.Id))
             return false;
-        _context.Customers.Update(customer);
-        _context.SaveChanges();
+        context.Customers.Update(customer);
+        context.SaveChanges();
         return true;
     }
 
@@ -45,12 +38,12 @@ public class CustomerRepository : ICustomerRepository
     {
         if (!CustomerExists(id))
             return false;
-        var toRemove = _context.Customers.Find(id);
-        _context.Customers.Remove(toRemove);
-        _context.SaveChanges();
+        var toRemove = context.Customers.Find(id);
+        context.Customers.Remove(toRemove);
+        context.SaveChanges();
         return true;
     }
 
     public List<Customer> GetBySupportRepId(int id) =>
-        _context.Customers.Where(a => a.SupportRepId == id).AsNoTrackingWithIdentityResolution().ToList();
+        context.Customers.Where(a => a.SupportRepId == id).AsNoTrackingWithIdentityResolution().ToList();
 }
