@@ -1,23 +1,11 @@
-﻿using Chinook.Domain.Converters;
-using Chinook.Domain.Entities;
+﻿namespace Chinook.Domain.ApiModels;
 
-namespace Chinook.Domain.ApiModels;
-
-public class AlbumApiModel : BaseApiModel, IConvertModel<Album>
+public class AlbumApiModel : BaseApiModel
 {
     public string Title { get; set; }
     public int ArtistId { get; set; }
-    public string ArtistName { get; set; }
 
     public ArtistApiModel Artist { get; set; }
 
-    public List<TrackApiModel> Tracks { get; set; }
-
-    public Album Convert() =>
-        new()
-        {
-            Id = Id,
-            ArtistId = ArtistId,
-            Title = Title ?? string.Empty
-        };
+    public IList<TrackApiModel> Tracks { get; set; }
 }
