@@ -1,9 +1,6 @@
-﻿using Chinook.Domain.Converters;
-using Chinook.Domain.Entities;
+﻿namespace Chinook.Domain.ApiModels;
 
-namespace Chinook.Domain.ApiModels;
-
-public class CustomerApiModel : BaseApiModel, IConvertModel<Customer>
+public class CustomerApiModel : BaseApiModel
 {
     public string? FirstName { get; set; }
     public string? LastName { get; set; }
@@ -17,28 +14,8 @@ public class CustomerApiModel : BaseApiModel, IConvertModel<Customer>
     public string? Fax { get; set; }
     public string? Email { get; set; }
     public int SupportRepId { get; set; }
-    public string? SupportRepName { get; set; }
 
     public IList<InvoiceApiModel>? Invoices { get; set; }
 
     public EmployeeApiModel? SupportRep { get; set; }
-
-
-    public Customer Convert() =>
-        new()
-        {
-            Id = Id,
-            FirstName = FirstName,
-            LastName = LastName,
-            Company = Company,
-            Address = Address,
-            City = City,
-            State = State,
-            Country = Country,
-            PostalCode = PostalCode,
-            Phone = Phone,
-            Fax = Fax,
-            Email = Email,
-            SupportRepId = SupportRepId
-        };
 }
