@@ -1,18 +1,8 @@
-﻿using Chinook.Domain.Converters;
-using Chinook.Domain.Entities;
+﻿namespace Chinook.Domain.ApiModels;
 
-namespace Chinook.Domain.ApiModels;
-
-public class MediaTypeApiModel : BaseApiModel, IConvertModel<MediaType>
+public class MediaTypeApiModel : BaseApiModel
 {
     public string? Name { get; set; }
 
     public IList<TrackApiModel>? Tracks { get; set; }
-
-    public MediaType Convert() =>
-        new()
-        {
-            Id = Id,
-            Name = Name
-        };
 }
