@@ -3,20 +3,13 @@ using Chinook.Domain.Helpers;
 
 namespace Chinook.Domain.Enrichers;
 
-public class EmployeesEnricher : ListEnricher<List<EmployeeApiModel>>
+public class EmployeesEnricher(EmployeeEnricher enricher) : ListEnricher<List<EmployeeApiModel>>
 {
-    private readonly EmployeeEnricher _enricher;
-
-    public EmployeesEnricher(EmployeeEnricher enricher)
-    {
-        _enricher = enricher;
-    }
-
     public override async Task Process(object representations)
     {
         foreach (var employee in (IEnumerable<EmployeeApiModel>)representations)
         {
-            await _enricher.Process(employee as EmployeeApiModel);
+            await enricher.Process(employee as EmployeeApiModel);
         }
     }
 }
