@@ -5,22 +5,13 @@ using Microsoft.AspNetCore.Routing;
 
 namespace Chinook.Domain.Enrichers;
 
-public class GenreEnricher : Enricher<GenreApiModel>
+public class GenreEnricher(IHttpContextAccessor accessor, LinkGenerator linkGenerator) : Enricher<GenreApiModel>
 {
-    private readonly IHttpContextAccessor _accessor;
-    private readonly LinkGenerator _linkGenerator;
-
-    public GenreEnricher(IHttpContextAccessor accessor, LinkGenerator linkGenerator)
-    {
-        _accessor = accessor;
-        _linkGenerator = linkGenerator;
-    }
-    
     public override Task Process(GenreApiModel? representation)
     {
-        var httpContext = _accessor.HttpContext;
+        var httpContext = accessor.HttpContext;
 
-        var url = _linkGenerator.GetUriByName(
+        var url = linkGenerator.GetUriByName(
             httpContext!,
             "GetGenreById",
             new { id = representation.Id },
@@ -33,6 +24,24 @@ public class GenreEnricher : Enricher<GenreApiModel>
             Title = $"genre: #{representation.Id}",
             Href = url!
         });
+        
+        // enrich the tracks
+        foreach (var track in representation.Tracks)
+        {
+            var urlTrack = linkGenerator.GetUriByName(
+                httpContext,
+                "GetTrackById",
+                new { id = track.Id },
+                scheme: "https"
+            );
+
+            track.AddLink(new Link
+            {
+                Rel = track.Id.ToString(),
+                Title = $"Track: #{track.Id}",
+                Href = urlTrack
+            });
+        }
 
         return Task.CompletedTask;
     }
