@@ -5,22 +5,13 @@ using Microsoft.AspNetCore.Routing;
 
 namespace Chinook.Domain.Enrichers;
 
-public class InvoiceLineEnricher : Enricher<InvoiceLineApiModel>
+public class InvoiceLineEnricher(IHttpContextAccessor accessor, LinkGenerator linkGenerator) : Enricher<InvoiceLineApiModel>
 {
-    private readonly IHttpContextAccessor _accessor;
-    private readonly LinkGenerator _linkGenerator;
-
-    public InvoiceLineEnricher(IHttpContextAccessor accessor, LinkGenerator linkGenerator)
-    {
-        _accessor = accessor;
-        _linkGenerator = linkGenerator;
-    }
-    
     public override Task Process(InvoiceLineApiModel? representation)
     {
-        var httpContext = _accessor.HttpContext;
+        var httpContext = accessor.HttpContext;
 
-        var url = _linkGenerator.GetUriByName(
+        var url = linkGenerator.GetUriByName(
             httpContext!,
             "GetInvoiceLineById",
             new { id = representation.Id },
@@ -32,6 +23,36 @@ public class InvoiceLineEnricher : Enricher<InvoiceLineApiModel>
             Rel = representation.Id.ToString(),
             Title = $"InvoiceLine: #{representation.Id}",
             Href = url!
+        });
+        
+        // enrich invoice
+        var urlInvoice = linkGenerator.GetUriByName(
+            httpContext,
+            "GetInvoiceById",
+            new { id = representation.InvoiceId },
+            scheme: "https"
+        );
+
+        representation.Invoice.AddLink(new Link
+        {
+            Rel = representation.InvoiceId.ToString(),
+            Title = $"Invoice: #{representation.InvoiceId}",
+            Href = urlInvoice
+        });
+        
+        //enrich track
+        var urlTrack = linkGenerator.GetUriByName(
+            httpContext,
+            "GetTrackById",
+            new { id = representation.TrackId },
+            scheme: "https"
+        );
+
+        representation.Track.AddLink(new Link
+        {
+            Rel = representation.TrackId.ToString(),
+            Title = $"Track: #{representation.TrackId}",
+            Href = urlTrack
         });
 
         return Task.CompletedTask;
