@@ -3,20 +3,13 @@ using Chinook.Domain.Helpers;
 
 namespace Chinook.Domain.Enrichers;
 
-public class InvoicesEnricher : ListEnricher<List<InvoiceApiModel>>
+public class InvoicesEnricher(InvoiceEnricher enricher) : ListEnricher<List<InvoiceApiModel>>
 {
-    private readonly InvoiceEnricher _enricher;
-
-    public InvoicesEnricher(InvoiceEnricher enricher)
-    {
-        _enricher = enricher;
-    }
-
     public override async Task Process(object representations)
     {
         foreach (var invoice in (IEnumerable<InvoiceApiModel>)representations)
         {
-            await _enricher.Process(invoice as InvoiceApiModel);
+            await enricher.Process(invoice as InvoiceApiModel);
         }
     }
 }
