@@ -3,20 +3,13 @@ using Chinook.Domain.Helpers;
 
 namespace Chinook.Domain.Enrichers;
 
-public class TracksEnricher : ListEnricher<List<TrackApiModel>>
+public class TracksEnricher(TrackEnricher enricher) : ListEnricher<List<TrackApiModel>>
 {
-    private readonly TrackEnricher _enricher;
-
-    public TracksEnricher(TrackEnricher enricher)
-    {
-        _enricher = enricher;
-    }
-
     public override async Task Process(object representations)
     {
         foreach (var track in (IEnumerable<TrackApiModel>)representations)
         {
-            await _enricher.Process(track as TrackApiModel);
+            await enricher.Process(track as TrackApiModel);
         }
     }
 }
