@@ -1,9 +1,6 @@
-﻿using Chinook.Domain.ApiModels;
-using Chinook.Domain.Converters;
+﻿namespace Chinook.Domain.Entities;
 
-namespace Chinook.Domain.Entities;
-
-public partial class Album : IConvertModel<AlbumApiModel>
+public partial class Album
 {
     public Album()
     {
@@ -16,12 +13,4 @@ public partial class Album : IConvertModel<AlbumApiModel>
 
     public virtual Artist Artist { get; set; } = null!;
     public virtual ICollection<Track> Tracks { get; set; }
-
-    public AlbumApiModel Convert() =>
-        new()
-        {
-            Id = Id,
-            ArtistId = ArtistId,
-            Title = Title
-        };
 }
