@@ -1,9 +1,6 @@
-﻿using Chinook.Domain.ApiModels;
-using Chinook.Domain.Converters;
+﻿namespace Chinook.Domain.Entities;
 
-namespace Chinook.Domain.Entities;
-
-public sealed class Customer : IConvertModel<CustomerApiModel>
+public sealed class Customer
 {
     public Customer()
     {
@@ -23,27 +20,6 @@ public sealed class Customer : IConvertModel<CustomerApiModel>
     public string? Fax { get; set; }
     public string? Email { get; set; }
     public int SupportRepId { get; set; }
-
-
     public Employee? SupportRep { get; set; }
-
     public ICollection<Invoice>? Invoices { get; set; }
-
-    public CustomerApiModel Convert() =>
-        new()
-        {
-            Id = Id,
-            FirstName = FirstName,
-            LastName = LastName,
-            Company = Company,
-            Address = Address,
-            City = City,
-            State = State,
-            Country = Country,
-            PostalCode = PostalCode,
-            Phone = Phone,
-            Fax = Fax,
-            Email = Email,
-            SupportRepId = SupportRepId
-        };
 }
