@@ -1,9 +1,6 @@
-﻿using Chinook.Domain.ApiModels;
-using Chinook.Domain.Converters;
+﻿namespace Chinook.Domain.Entities;
 
-namespace Chinook.Domain.Entities;
-
-public sealed class Genre : IConvertModel<GenreApiModel>
+public sealed class Genre
 {
     public Genre()
     {
@@ -13,11 +10,4 @@ public sealed class Genre : IConvertModel<GenreApiModel>
     public int Id { get; set; }
     public string? Name { get; set; }
     public ICollection<Track>? Tracks { get; set; }
-
-    public GenreApiModel Convert() =>
-        new()
-        {
-            Id = Id,
-            Name = Name
-        };
 }
