@@ -1,9 +1,6 @@
-﻿using Chinook.Domain.ApiModels;
-using Chinook.Domain.Converters;
+﻿namespace Chinook.Domain.Entities;
 
-namespace Chinook.Domain.Entities;
-
-public sealed class MediaType : IConvertModel<MediaTypeApiModel>
+public sealed class MediaType
 {
     public MediaType()
     {
@@ -14,11 +11,4 @@ public sealed class MediaType : IConvertModel<MediaTypeApiModel>
     public string? Name { get; set; }
 
     public ICollection<Track>? Tracks { get; set; }
-
-    public MediaTypeApiModel Convert() =>
-        new()
-        {
-            Id = Id,
-            Name = Name
-        };
 }
