@@ -1,8 +1,6 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿namespace Chinook.Domain.Helpers;
 
-namespace Chinook.Domain.Helpers;
-
-public abstract class ListEnricher<T> : IListEnricher
+public abstract class ListEnricher<T> : IEnricher
 {
     public virtual Task<bool> Match(object target) =>
         target switch
