@@ -1,8 +1,6 @@
 using Chinook.Domain.ApiModels;
 using Chinook.Domain.Entities;
-using Chinook.Domain.Extensions;
 using FluentValidation;
-using Microsoft.Extensions.Caching.Memory;
 
 namespace Chinook.Domain.Supervisor;
 
@@ -10,52 +8,25 @@ public partial class ChinookSupervisor
 {
     public List<ArtistApiModel> GetAllArtist()
     {
-        List<Artist> artists = _artistRepository.GetAll();
-        var artistApiModels = artists.ConvertAll();
+        var artists = artistRepository.GetAll().Select(mapper.Map<ArtistApiModel>).ToList();
 
-        foreach (var artist in artistApiModels)
-        {
-            var cacheEntryOptions =
-                new MemoryCacheEntryOptions().SetSlidingExpiration(TimeSpan.FromSeconds(604800))
-                    .AbsoluteExpirationRelativeToNow = TimeSpan.FromSeconds(604800);
-            ;
-            _cache.Set(string.Concat("Artist-", artist.Id), artist, (TimeSpan)cacheEntryOptions);
-        }
-
-        return artistApiModels;
+        return artists;
     }
 
     public ArtistApiModel GetArtistById(int id)
     {
-        var artistApiModelCached = _cache.Get<ArtistApiModel>(string.Concat("Artist-", id));
+        var artist = artistRepository.GetById(id);
+        var artistApiModel = _mapper.Map<ArtistApiModel>(artist);
 
-        if (artistApiModelCached != null)
-        {
-            return artistApiModelCached;
-        }
-        else
-        {
-            var artist = _artistRepository.GetById(id);
-            var artistApiModel = artist.Convert();
-            artistApiModel.Albums = (_albumRepository.GetByArtistId(artist.Id)).ConvertAll().ToList();
-
-            var cacheEntryOptions =
-                new MemoryCacheEntryOptions().SetSlidingExpiration(TimeSpan.FromSeconds(604800))
-                    .AbsoluteExpirationRelativeToNow = TimeSpan.FromSeconds(604800);
-            ;
-            _cache.Set(string.Concat("Artist-", artistApiModel.Id), artistApiModel, (TimeSpan)cacheEntryOptions);
-
-            return artistApiModel;
-        }
+        return artistApiModel;
     }
 
     public ArtistApiModel AddArtist(ArtistApiModel newArtistApiModel)
     {
         _artistValidator.ValidateAndThrowAsync(newArtistApiModel);
-
-        var artist = newArtistApiModel.Convert();
-
-        artist = _artistRepository.Add(artist);
+        var artist = _mapper.Map<Artist>(newArtistApiModel);
+        
+        artist = artistRepository.Add(artist);
         newArtistApiModel.Id = artist.Id;
         return newArtistApiModel;
     }
@@ -63,15 +34,11 @@ public partial class ChinookSupervisor
     public bool UpdateArtist(ArtistApiModel artistApiModel)
     {
         _artistValidator.ValidateAndThrowAsync(artistApiModel);
+        var artist = _mapper.Map<Artist>(artistApiModel);
 
-        var artist = _artistRepository.GetById(artistApiModel.Id);
-        
-        artist.Id = artistApiModel.Id;
-        artist.Name = artistApiModel.Name ?? string.Empty;
-
-        return _artistRepository.Update(artist);
+        return artistRepository.Update(artist);
     }
 
     public bool DeleteArtist(int id)
-        => _artistRepository.Delete(id);
+        => artistRepository.Delete(id);
 }
