@@ -1,8 +1,6 @@
 using Chinook.Domain.ApiModels;
 using Chinook.Domain.Entities;
-using Chinook.Domain.Extensions;
 using FluentValidation;
-using Microsoft.Extensions.Caching.Memory;
 
 namespace Chinook.Domain.Supervisor;
 
@@ -10,51 +8,23 @@ public partial class ChinookSupervisor
 {
     public List<GenreApiModel> GetAllGenre()
     {
-        List<Genre> genres = _genreRepository.GetAll();
-        var genreApiModels = genres.ConvertAll();
+        var genres = genreRepository.GetAll().Select(mapper.Map<GenreApiModel>).ToList();
 
-        foreach (var genre in genreApiModels)
-        {
-            var cacheEntryOptions =
-                new MemoryCacheEntryOptions().SetSlidingExpiration(TimeSpan.FromSeconds(604800))
-                    .AbsoluteExpirationRelativeToNow = TimeSpan.FromSeconds(604800);
-            ;
-            _cache.Set(string.Concat("Genre-", genre.Id), genre, (TimeSpan)cacheEntryOptions);
-        }
-
-        return genreApiModels;
+        return genres;
     }
 
     public GenreApiModel? GetGenreById(int id)
     {
-        var genreApiModelCached = _cache.Get<GenreApiModel>(string.Concat("Genre-", id));
+        var genre = _genreRepository.GetById(id);
+        var genreApiModel = _mapper.Map<GenreApiModel>(genre);
 
-        if (genreApiModelCached != null)
-        {
-            return genreApiModelCached;
-        }
-        else
-        {
-            var genre = _genreRepository.GetById(id);
-            var genreApiModel = genre.Convert();
-            genreApiModel.Tracks = (GetTrackByGenreId(genreApiModel.Id)).ToList();
-
-            var cacheEntryOptions =
-                new MemoryCacheEntryOptions().SetSlidingExpiration(TimeSpan.FromSeconds(604800))
-                    .AbsoluteExpirationRelativeToNow = TimeSpan.FromSeconds(604800);
-            ;
-            _cache.Set(string.Concat("Genre-", genreApiModel.Id), genreApiModel, (TimeSpan)cacheEntryOptions);
-
-            return genreApiModel;
-        }
+        return genreApiModel;
     }
 
     public GenreApiModel AddGenre(GenreApiModel newGenreApiModel)
     {
         _genreValidator.ValidateAndThrowAsync(newGenreApiModel);
-
-        var genre = newGenreApiModel.Convert();
-
+        var genre = _mapper.Map<Genre>(newGenreApiModel);
         genre = _genreRepository.Add(genre);
         newGenreApiModel.Id = genre.Id;
         return newGenreApiModel;
@@ -63,12 +33,7 @@ public partial class ChinookSupervisor
     public bool UpdateGenre(GenreApiModel genreApiModel)
     {
         _genreValidator.ValidateAndThrowAsync(genreApiModel);
-
-        var genre = _genreRepository.GetById(genreApiModel.Id);
-        
-        genre.Id = genreApiModel.Id;
-        genre.Name = genreApiModel.Name ?? string.Empty;
-
+        var genre = _mapper.Map<Genre>(genreApiModel);
         return _genreRepository.Update(genre);
     }
 
