@@ -1,8 +1,6 @@
 using Chinook.Domain.ApiModels;
 using Chinook.Domain.Entities;
-using Chinook.Domain.Extensions;
 using FluentValidation;
-using Microsoft.Extensions.Caching.Memory;
 
 namespace Chinook.Domain.Supervisor;
 
@@ -10,65 +8,24 @@ public partial class ChinookSupervisor
 {
     public List<InvoiceLineApiModel> GetAllInvoiceLine()
     {
-        List<InvoiceLine> invoiceLines = _invoiceLineRepository.GetAll();
-        var invoiceLineApiModels = invoiceLines.ConvertAll();
+        var invoiceLines = _invoiceLineRepository.GetAll().Select(mapper.Map<InvoiceLineApiModel>).ToList();
 
-        foreach (var invoiceLine in invoiceLineApiModels)
-        {
-            var cacheEntryOptions =
-                new MemoryCacheEntryOptions().SetSlidingExpiration(TimeSpan.FromSeconds(604800))
-                    .AbsoluteExpirationRelativeToNow = TimeSpan.FromSeconds(604800);
-            ;
-            _cache.Set(string.Concat("InvoiceLine-", invoiceLine.Id), invoiceLine, (TimeSpan)cacheEntryOptions);
-        }
-
-        return invoiceLineApiModels;
+        return invoiceLines;
     }
 
     public InvoiceLineApiModel GetInvoiceLineById(int id)
     {
-        var invoiceLineApiModelCached = _cache.Get<InvoiceLineApiModel>(string.Concat("InvoiceLine-", id));
+        var invoiceLine = _invoiceLineRepository.GetById(id);
+        var invoiceLineApiModel = _mapper.Map<InvoiceLineApiModel>(invoiceLine);
 
-        if (invoiceLineApiModelCached != null)
-        {
-            return invoiceLineApiModelCached;
-        }
-        else
-        {
-            var invoiceLine = _invoiceLineRepository.GetById(id);
-            var invoiceLineApiModel = invoiceLine.Convert();
-            invoiceLineApiModel.Track = GetTrackById(invoiceLineApiModel.TrackId);
-            invoiceLineApiModel.Invoice = GetInvoiceById(invoiceLineApiModel.InvoiceId);
-            if (invoiceLineApiModel.Track != null) invoiceLineApiModel.TrackName = invoiceLineApiModel.Track.Name;
-
-            var cacheEntryOptions =
-                new MemoryCacheEntryOptions().SetSlidingExpiration(TimeSpan.FromSeconds(604800))
-                    .AbsoluteExpirationRelativeToNow = TimeSpan.FromSeconds(604800);
-            ;
-            _cache.Set(string.Concat("InvoiceLine-", invoiceLineApiModel.Id), invoiceLineApiModel,
-                (TimeSpan)cacheEntryOptions);
-
-            return invoiceLineApiModel;
-        }
-    }
-
-    public List<InvoiceLineApiModel> GetInvoiceLineByInvoiceId(int id)
-    {
-        var invoiceLines = _invoiceLineRepository.GetByInvoiceId(id);
-        return invoiceLines.ConvertAll();
-    }
-
-    public List<InvoiceLineApiModel> GetInvoiceLineByTrackId(int id)
-    {
-        var invoiceLines = _invoiceLineRepository.GetByTrackId(id);
-        return invoiceLines.ConvertAll();
+        return invoiceLineApiModel;
     }
 
     public InvoiceLineApiModel AddInvoiceLine(InvoiceLineApiModel newInvoiceLineApiModel)
     {
         _invoiceLineValidator.ValidateAndThrowAsync(newInvoiceLineApiModel);
-
-        var invoiceLine = newInvoiceLineApiModel.Convert();
+        
+        var invoiceLine = _mapper.Map<InvoiceLine>(newInvoiceLineApiModel);
 
         invoiceLine = _invoiceLineRepository.Add(invoiceLine);
         newInvoiceLineApiModel.Id = invoiceLine.Id;
@@ -79,17 +36,23 @@ public partial class ChinookSupervisor
     {
         _invoiceLineValidator.ValidateAndThrowAsync(invoiceLineApiModel);
 
-        var invoiceLine = _invoiceLineRepository.GetById(invoiceLineApiModel.InvoiceId);
-        
-        invoiceLine.Id = invoiceLineApiModel.Id;
-        invoiceLine.InvoiceId = invoiceLineApiModel.InvoiceId;
-        invoiceLine.TrackId = invoiceLineApiModel.TrackId;
-        invoiceLine.UnitPrice = invoiceLineApiModel.UnitPrice;
-        invoiceLine.Quantity = invoiceLineApiModel.Quantity;
+        var invoiceLine = _mapper.Map<InvoiceLine>(invoiceLineApiModel);
 
         return _invoiceLineRepository.Update(invoiceLine);
     }
 
     public bool DeleteInvoiceLine(int id)
         => _invoiceLineRepository.Delete(id);
+    
+    public List<InvoiceLineApiModel> GetInvoiceLineByInvoiceId(int id)
+    {
+        var invoiceLines = _invoiceLineRepository.GetByInvoiceId(id).Select(mapper.Map<InvoiceLineApiModel>).ToList();
+        return invoiceLines;
+    }
+
+    public List<InvoiceLineApiModel> GetInvoiceLineByTrackId(int id)
+    {
+        var invoiceLines = _invoiceLineRepository.GetByTrackId(id).Select(mapper.Map<InvoiceLineApiModel>).ToList();
+        return invoiceLines;
+    }
 }
